MathsForKids
Maths for Kids Pro is the best app so far in this category which will
provide your kids the easy and fun way of learning maths. Learning maths
is no more a difficulty for your kids. This app provides an easy
interface which allows your kids to practice all basic mathematics. It
includes addition, subtraction, multiplication and puzzles like number
guessing and more fun games